Demo

+++ Plot/Data/Models/Fixtures/UpdateFixtureInstance.cs
@@ -20,11 +20,11 @@ public record UpdateFixtureInstance
     public int? TUID { get; set; }
     public int? FIXTURE_TUID { get; set; }
     public int? FLOORSET_TUID { get; set; }
-    public double? X_POS { get; set; }
-    public double? Y_POS { get; set; }
-    public double? ALLOCATED_LF { get; set; }
+    public int? X_POS { get; set; }
+    public int? Y_POS { get; set; }
+    public int? ALLOCATED_LF { get; set; }
     public int? HANGER_STACK { get; set; }
-    public float? TOT_LF { get; set; }
+    public int? TOT_LF { get; set; }
     public string? CATEGORY { get; set; }
     public string? NOTE { get; set; }
 }
